More Teeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeests

+++ IPoAT-Core/src/main/java/dhbw/ipoat/player/Player.java
@@ -89,7 +89,7 @@ public class Player {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(float money) {
         this.money = money;
     }
